refactor: split universeBuilder into focused modules behind thin barrel (#2529)
Split the ~2000-line universeBuilder.js god-module along its natural
seams into sibling modules (sanitize / storeFacade / crud / sync /
compile) and reduce the original path to a thin re-export barrel so
all 94 existing importers keep working. Pure structural refactor with
no behavior change; all existing suites (incl. DB-backed) stay green.

catalogDB.js split deferred to remaining work under the same issue.
